+++ src/features/manageProfileSettings/utils/index.ts
@@ -1,2 +1,3 @@
 export { FilteredPointerSensor, FilteredKeyboardSensor } from './filteredSensors';
 export { componentFromId, getProfileChildren, getSettingsChildren, childrenDifference } from './children';
+export { chooseModifiers } from './modifiers';